LMain screen layout change, map functionality change
