+++ front-end/package.json
@@ -54,6 +54,7 @@
     "react-leaflet": "2.2.0",
     "react-leaflet-geojson-patterns": "1.1.0",
     "react-router-dom": "5.2.0",
+    "react-virtualized-auto-sizer": "^1.0.6",
     "recharts": "1.4.2",
     "typeface-roboto": "0.0.54",
     "yarn": "^1.22.10"
